Beautify and better examples
The JSX example was poor due to its misuse of `do`. `do`, in this example, was being used where a ternary is often used, so to ensure the reason to use `do` was more effectively communicated I designed a more complicated example where a ternary would be unpreferable. Semicolons were being used off and on, therefore, I removed them as to avoid confusion. Increased readability in the conditional section and removed unnecessary curly braces.